utc time conversion for backend
